Refactor the document actions component
